+++ src/components/PartCollector/CurrentView.ts
@@ -1,0 +1,4 @@
+export enum CurrentView {
+  ADD_PARTS = 'Add Parts',
+  VIEW_PARTS = 'View Parts'
+}