+++ src/model/partCollector/Color.ts
@@ -1,0 +1,6 @@
+export interface Color {
+  id: string;
+  description: string;
+  rgb: string;
+  translucent: boolean;
+}