+++ src/model/partCollector/PartDisplay.ts
@@ -1,0 +1,9 @@
+import { Part } from "./Part";
+
+export interface PartDisplay {
+  key: string;
+  set: string;
+  quantity: number;
+  comment: string;
+  part: Part;
+}