+++ src/model/partCollector/PartsResponse.ts
@@ -1,0 +1,8 @@
+import { PartResult } from "./PartResult";
+
+export interface PartsResponse {
+  count: number;
+  next?: string;
+  previous?: string;
+  results: PartResult[];
+}